claude-agent-sdk-in-action/code-examples/chapter-02/requirements.txt
anthropic==0.34.0
python-dotenv==1.0.0
pyyaml==6.0.1
matplotlib==3.8.0
pandas==2.1.0
reportlab==4.0.5
